# Backend-aspnet/MockServer/AiFastApi/Models/SendBroadcastRequest.cs
namespace MockServer.AiFastApi.Models
{

    public class SendBroadcastRequest
    {
        public List<string> Phonenumbers { get; set; }
        public string Broadcastmessage { get; set; }
    }


}